refactoring execute, updated tests
