+++ test/__init__.py
@@ -1,0 +1,38 @@
+import unittest
+
+class TestCase(unittest.TestCase):
+    """
+    Base class for Otter unit and integration tests that includes some helpful logging and custom 
+    assert methods.
+    """
+
+    def setUp(self):
+        print("\n" + ("=" * 80) + f"\nRunning {self.id()}\n" + ("=" * 80))
+        return super().setUp()
+
+    def assertFilesEqual(self, p1, p2):
+        try:
+            with open(p1) as f1:
+                with open(p2) as f2:
+                    self.assertEqual(f1.read(), f2.read(), f"Contents of {p1} did not equal contents of {p2}")
+        
+        except UnicodeDecodeError:
+            with open(p1, "rb") as f1:
+                with open(p2, "rb") as f2:
+                    self.assertEqual(f1.read(), f2.read(), f"Contents of {p1} did not equal contents of {p2}")
+
+    def assertDirsEqual(self, dir1, dir2, ignore_ext=[]):
+        self.assertTrue(os.path.exists(dir1), f"{dir1} does not exist")
+        self.assertTrue(os.path.exists(dir2), f"{dir2} does not exist")
+        self.assertTrue(os.path.isfile(dir1) == os.path.isfile(dir2), f"{dir1} and {dir2} have different type")
+
+        if os.path.isfile(dir1):
+            if os.path.splitext(dir1)[1] not in ignore_ext:
+                self.assertFilesEqual(dir1, dir2)
+
+        else:
+            self.assertEqual(os.listdir(dir1), os.listdir(dir2), f"{dir1} and {dir2} have different contents")
+            for f1, f2 in zip(os.listdir(dir1), os.listdir(dir2)):
+                f1, f2 = os.path.join(dir1, f1), os.path.join(dir2, f2)
+                self.assertDirsEqual(f1, f2, ignore_ext=ignore_ext)
+
